backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.1
langchain-postgres==0.0.12
psycopg2-binary==2.9.9
pgvector==0.3.3
langdetect==1.0.9
python-dotenv==1.0.1
pydantic==2.9.2
websockets==13.1
pypdf==4.3.1
python-multipart==0.0.12